la til interfaces til helperklassene utenom attack. Informasjon om attack vil ligge i coordinates i databasen og vi kan bruke attack feltet i Coordinate klassen.
Passhash lagt i egen klasse. Når en bruker logger inn i gui skal ikke all info i farmerobjektet sendes fra server til client. 
Kun selve passordet må sendes. Client sjekker om passordet stemmer og sender ny request for å få ett farmerobjekt. 
Passhash lagt til i farmer slik at passord kan forandres. 

Feature (could have)
Hvis bonden har glemt passord og vi har implementert kryptering, må server sende link i epost hvor brukeren lager ett nytt passord. 
Nytt passord sendes til server via Passhash objekt.
