+++ sheepfarmer3000/src/main/java/org/prosjekt/helperclasses/Farmer.java
@@ -1,148 +1,49 @@
+/*
+ * To change this template, choose Tools | Templates
+ * and open the template in the editor.
+ */
 package org.prosjekt.helperclasses;
 
-import java.io.Serializable;
+import org.prosjekt.helperclasses.impl.SheepImpl;
 import java.util.ArrayList;
+import java.util.List;
 
-public class Farmer implements Serializable,IFarmer {
-    private final int id; //brukes som brukernavn
-    private String firstName;
-    private String lastName;
-    private ArrayList<Sheep> sheeps;
-    private String passhash;
-    private String email;
-    private String phone;
-    //kontaktinfo til helper
-    private String helperFirstname;
-    private String helperLastName;
-    private String helperPhone;
-    private String helperEmail;
-    private ArrayList<Coordinate> cordinates; //koordinatene til gaaren/teigen
+/**
+ *
+ * @author Christoffer Buvik christofferbuvik@gmail.com
+ */
+public interface Farmer {
+    public int getId();
+    public Passhash getPasshash();
+    public void setPasshash(Passhash passhash);
+    
+    public String getEmail();
+    public void setEmail(String email);
+    
+    public String getPhone();
+    public void setPhone(String phone);
 
-    /*
-     *  Temporary constructor, used under db development. 
-     */
-    public Farmer(int id) {
-        this.id = id;
-        
-        
-        
-    }
-    
-    
-    
-    public Farmer(int id, String firstName, String lastName, String password, String email, String phone, String helperFirstName, String helperLastName, String helperPhone, String helperEmail) {
-        
-        this.id = id;
-        this.firstName = firstName;
-        this.lastName = lastName;
-        this.passhash = password;
-        this.email = email;
-        this.phone = phone;
-        this.helperFirstname = helperFirstName;
-        this.helperLastName = helperLastName;
-        this.helperPhone = helperPhone;
-        this.helperEmail = helperEmail;
-        this.sheeps = new ArrayList<Sheep>();
-    }
-    
- 
-    
-    public String getEmail() {
-        return email;
-    }
-    
-    public void setEmail(String email) {
-        this.email = email;
-    }
-    
-    public String getPhone() {
-        return phone;
-    }
-    
-    public void setPhone(String phone) {
-        this.phone = phone;
-    }
+    public String getHelperFirstname();
+    public void setHelperFirstname(String helperFirstname);
 
-    public String getHelperFirstname() {
-        return helperFirstname;
-    }
-
-    public void setHelperFirstname(String helperFirstname) {
-        this.helperFirstname = helperFirstname;
-    }
-
-    public String getHelperLastName() {
-        return helperLastName;
-    }
-
-    public void setHelperLastName(String helperLastName) {
-        this.helperLastName = helperLastName;
-    }
+    public String getHelperLastName();
+    public void setHelperLastName(String helperLastName);
     
-    public String getHelperPhone() {
-        return helperPhone;
-    }
+    public String getHelperPhone();    
+    public void setHelperPhone(String helperPhone);
     
-    public void setHelperPhone(String helperPhone) {
-        this.helperPhone = helperPhone;
-    }
+    public String getHelperEmail();
+    public void setHelperEmail(String helperEmail);
     
-    public String getHelperEmail() {
-        return helperEmail;
-    }
     
-    public void setHelperEmail(String helperEmail) {
-        this.helperEmail = helperEmail;
-    }
+    public String getFirstName();
+    public void setFirstName(String firstname);
     
-    public int getId() {
-        return id;
-    }
+    public String getLastName();
+    public void setLastName(String lastname);
     
-    public String getFirstName() {
-        return firstName;
-    }
-    
-    public String getLastName() {
-        return lastName;
-    }
-    
-    public ArrayList<Sheep> getSheeps() {
-        return sheeps;
-    }
-
-    public void setFirstName(String firstName) {
-        this.firstName = firstName;
-    }
-
-    public void setLastName(String lastName) {
-        this.lastName = lastName;
-    }
-
-    public void setSheeps(ArrayList<Sheep> sheeps) {
-        this.sheeps = sheeps;
-    }
-
-    public String getPasshash() {
-        return passhash;
-    }
-
-    public void setPasshash(String passhash) {
-        this.passhash = passhash;
-    }
+    public void addSheep(Sheep sheep);
+    public List<Sheep> getSheeps();
 
     
-
-    public void setCordinates(ArrayList<Coordinate> coordinates) {
-        this.cordinates = coordinates;
-    }
-
-    @Override
-    public String toString() {
-        return "Farmer{" + "id=" + id + ", firstName=" + firstName + ", lastName=" + lastName + ", password=" + passhash + ", email=" + email + ", phone=" + phone + '}';
-    }
-    
-    
-
 }
-
