+++ sheepfarmer3000/src/main/java/org/prosjekt/helperclasses/Sheep.java
@@ -1,65 +1,28 @@
+/*
+ * To change this template, choose Tools | Templates
+ * and open the template in the editor.
+ */
 package org.prosjekt.helperclasses;
 
-import java.io.Serializable;
-import java.util.ArrayList;
-
+import java.util.List;
 import org.joda.time.DateTime;
 
-public class Sheep implements Serializable {
-	private final int id;
-	private boolean alive;
-	private int weight;
-	private String attack;
-	private final DateTime birth;
-	private ArrayList<Coordinate> cordinates;
-	private Farmer farmer;
-
-	public Sheep(int id, int weight, DateTime birth, Farmer farmer) {
-		this.id = id;
-		this.weight = weight;
-		this.birth = birth;
-		this.farmer = farmer;
-	}
-
-	public int getId() {
-		return this.id;
-	}
+/**
+ *
+ * @author Christoffer <christofferbuvik@gmail.com>
+ */
+public interface Sheep {
+        public int getId();
 	
-	public boolean getAlive() {
-		return this.alive;
-	}
+	public boolean getAlive();
+	public void setAlive(boolean status);
 	
-	public void setAlive(boolean status) {
-		this.alive = status;
-	}
+	public int getWeight();
+	public void setWeigth(int weight);
 	
-	public int getWeight() {
-		return this.weight;
-	}
-	
-	public void setWeigth(int weight) {
-		this.weight = weight;
-	}
-	
-	public DateTime getBirth() {
-		return this.birth;
-	}
-	
-	//Her må noe gjøres, jeg vet ikke hva Arraylisten skal inneholde
-	public ArrayList getAllCordinates() { 
-		return this.cordinates;
-	}
-	//Returnerer siste elementet i cordinates listen fra denne sauen.
-	public Coordinate getCurrentCordinates() {
-            //throw new UnsupportedOperationException("not yet implemented. ");
-		return this.cordinates.get(this.cordinates.size() - 1 );
-	}
-	
-	
-	//Trenger vi en setFarmer? bytte av eierskap?
-	public Farmer getFarmer() {
-		return this.farmer;
-	}
-	
-	
+	public DateTime getBirth();
+        
+        public List<Coordinate> getAllCoordinates();
+        public void addCoordinate(Coordinate coordinate);
+        public void setMostCurrentCoordinate(Coordinate coordinate);
 }
